+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format mostly follows [Keep a Changelog](http://keepachangelog.com/en/1.0.0/
   (PR#703, fixes #702, by Julien Palard)
 - Fix a ResourceWarning for unclosed files when running unit tests
   (PR#698, by Louis Sautier)
+- Add support for html2text 2.1.1 and newer by feature-checking `-utf8` support via `-help` (fixes #718)
 
 ## [2.25] -- 2022-03-15